Separate included specimens in lib and standalone build
I.e. the ReactSpecimen (and thus the huge babel-standalone) is not included in the standalone build
