+++ src/specimens-standalone.js
@@ -1,0 +1,23 @@
+import Audio from './specimens/Audio';
+import Code from './specimens/Code';
+import Color from './specimens/Color';
+import ColorPalette from './specimens/ColorPalette';
+import Html from './specimens/Html';
+import Hint from './specimens/Hint';
+import Image from './specimens/Image';
+import Type from './specimens/Type';
+import Download from './specimens/Download';
+import Video from './specimens/Video';
+
+export default {
+  audio: Audio,
+  code: Code,
+  color: Color,
+  'color-palette': ColorPalette,
+  html: Html,
+  hint: Hint,
+  image: Image,
+  type: Type,
+  download: Download,
+  video: Video
+};
